+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add way to specify contributor's role by adding `:role` key in hash syntax (see [bookspec.md](docs/bookspec.md#role))
 - add some docs about bookspec (see [bookspec.md](docs/bookspec.md))
+- add command `from-file` to create new Epuber project from existing EPUB file [#95](https://github.com/epuber-io/epuber/pull/95)
 
 ## v0.8.0 (2023-11-19)